compiler cleanup - move ParserErrorsToReasonStrings()
