--- src/StateSmith/Input/antlr4/Error.cs
@@ -1,59 +1,0 @@
-﻿using Antlr4.Runtime;
-
-#nullable enable
-
-namespace StateSmith.Input.antlr4;
-
-public abstract class Error
-{
-    public int line;
-    public int column;
-    public string message = "";
-    public RecognitionException exception;
-
-    public Error(int line, int column, string message, RecognitionException exception)
-    {
-        this.line = line;
-        this.column = column;
-        this.message = message;
-        this.exception = exception;
-    }
-
-    public abstract string BuildMessage();
-}
-
-public class ParseError : Error
-{
-    public IToken offendingSymbol;
-
-    public ParseError(IToken offendingSymbol, int line, int column, string message, RecognitionException exception) : base(line, column, message, exception)
-    {
-        this.offendingSymbol = offendingSymbol;
-    }
-
-    public override string BuildMessage()
-    {
-        var msg = $"{message} at line {line} column {column}.";
-        msg += $" Offending symbol: `{offendingSymbol.Text}`."; // todolow would be nice to print lexer's token name https://stackoverflow.com/questions/4403878/antlr-get-token-name
-
-        return msg;
-    }
-}
-
-public class LexerError : Error
-{
-    public int offendingChar;
-
-    public LexerError(int offendingChar, int line, int column, string message, RecognitionException exception) : base(line, column, message, exception)
-    {
-        this.offendingChar = offendingChar;
-    }
-
-    public override string BuildMessage()
-    {
-        var msg = $"{message} at line {line} column {column}.";
-        msg += $" Offending char: `{offendingChar}`."; // todolow would be nice to print lexer's token name https://stackoverflow.com/questions/4403878/antlr-get-token-name
-
-        return msg;
-    }
-}
