+++ src/StateSmith/Input/antlr4/LabelParser.cs
@@ -32,7 +32,7 @@ namespace StateSmith.Input.antlr4
             return errorListener.errors.Count > 0;
         }
 
-        public List<Error> GetErrors()
+        public List<AntlrError> GetErrors()
         {
             return errorListener.errors;
         }
